Safari-609.1.11.1/LayoutTests/fast/events/before-input-events-different-start-end-elements-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
Fired `onbeforeinput` handler!
Fired `oninput` handler!
Fired `onbeforeinput` handler!
Fired `oninput` handler!
Fired `onbeforeinput` handler!
Fired `oninput` handler!
Fired `onbeforeinput` handler!
Fired `oninput` handler!
Fired `onbeforeinput` handler!
Fired `oninput` handler!
Fired `onbeforeinput` handler!
Fired `oninput` handler!
Fired `onbeforeinput` handler!
Fired `oninput` handler!
Fired `onbeforeinput` handler!
Fired `oninput` handler!
Fired `onbeforeinput` handler!
Fired `oninput` handler!
Fired `onbeforeinput` handler!
Fired `oninput` handler!
Fired `onbeforeinput` handler!
Fired `oninput` handler!
